pyproject.toml
[project]
name = "RandomCatch"
version = "2.0"
description = "A Discord bot that allows users to catch math objects that randomly appear in designated channels."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord.py>=2.6.0",
]
